app/controllers/dashboard.py
"""
儀表板控制器模組
"""
from flask import Blueprint, render_template, jsonify
from flask_login import login_required, current_user
from google.analytics.data_v1beta import BetaAnalyticsDataClient
from google.analytics.data_v1beta.types import (
    RunReportRequest,
    DateRange,
    Metric,
    Dimension
)
from app import db

dashboard_bp = Blueprint('dashboard', __name__)

@dashboard_bp.route('/dashboard')
@login_required
def index():
    """
    儀表板首頁
    
    Returns:
        Response: 儀表板頁面
    """
    return render_template('dashboard/index.html')

@dashboard_bp.route('/api/dashboard/metrics')
@login_required
def get_metrics():
    """
    獲取儀表板指標數據
    
    Returns:
        Response: JSON 格式的指標數據
    """
    # TODO: 實現 Google Analytics 數據獲取
    return jsonify({
        'total_users': 1000,
        'active_users': 500,
        'conversion_rate': 0.05
    })

@dashboard_bp.route('/api/dashboard/daily-traffic-conversion')
@login_required
def daily_traffic_conversion():
    """
    獲取每日流量和轉化數據
    
    Returns:
        Response: JSON 格式的每日數據
    """
    # TODO: 實現 Google Analytics 數據獲取
    return jsonify({
        'dates': ['2024-01-01', '2024-01-02', '2024-01-03'],
        'traffic': [1000, 1200, 1100],
        'conversions': [50, 60, 55]
    }) 