requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
google-analytics-data==0.17.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pytest==8.0.2
SQLAlchemy==2.0.27
Werkzeug==3.0.1 